+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.1.2
 
 - add --version argument, used to print current otlet version
+- add exceptions.py and implement "custom" exceptions
+- add differentiation between when no package found in API and  when package version not found
 
 # 0.1.1